Update to stricter rate limits
